server/requirements.txt
fastapi
uvicorn[standard]
dashscope
python-dotenv
python-multipart
requests
